--- example/sigmod_demo/resources/log4j.properties
@@ -1,5 +1,0 @@
-log4j.rootLogger = INFO, stderr
-
-log4j.appender.stdout = org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.layout = org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%5p [%d] ({%t} %F[%M]:%L) - %m%n
