+++ example/sigmod_demo/resources/udf_list
@@ -1,6 +1,5 @@
 UDSF,store_embedding,UDFStoreEmbedding,EmbeddingUtils.py
-UDSF,get_embedding,UDFGetEmbedding,EmbeddingUtils.py
 UDSF,search_embedding,UDFSearchEmbedding,EmbeddingUtils.py
 UDSF,analyse_relation,UDFAnalyseRelation,EmbeddingUtils.py
+UDSF,merge,UDFMerge,EmbeddingUtils.py
 UDSF,ask_big_model,UDFAskBigModel,BigModelUtils.py
-UDSF,merge,UDFMerge,Merge.py
